Moved ContentWrapper styled-component to LogoGroup/styles.js
