+++ src/components/LogoGroup/styles.js
